changed input conversion from int to long

+++ deztoint/src/DecimalToBinaryWithoutComments.java
@@ -71,6 +71,6 @@ public class DecimalToBinaryWithoutComments
         {
             userInput = JOptionPane.showInputDialog("Bitte gültige Ganzzahl eingeben!");	
         }
-        return Integer.parseInt(userInput);
+        return Long.parseLong(userInput);
     }
 }
